Quality of Life and Bug Removal
Trevor, when you changed the size of everything you managed to somehow unhinge a bunch of stuff. I added pseudocode to the cup spoop, but also fixed a few unity-editor issues involving connections between certain objects and certain scripts. (For some reason the phone and the dialogue box were unhinged from things they needed to be attached to, coding-wise, but they should be good now).